--- a/datashader/spatial/__init__.py
+++ b/datashader/spatial/__init__.py
@@ -1,10 +1,10 @@
 from __future__ import absolute_import
 from datashader.spatial.points import read_parquet  # noqa (API import)
 
-from datashader.spatial.proximity import proximity  # noqa (API import)
-from datashader.spatial.proximity import great_circle_distance  # noqa (API import)
-from datashader.spatial.proximity import euclidean_distance  # noqa (API import)
-from datashader.spatial.proximity import manhattan_distance  # noqa (API import)
-
-from datashader.spatial.viewshed import viewshed  # noqa (API import)
-from datashader.spatial.zonal import zonal_stats  # noqa (API import)
+#from datashader.spatial.proximity import proximity  # noqa (API import)
+#from datashader.spatial.proximity import great_circle_distance  # noqa (API import)
+#from datashader.spatial.proximity import euclidean_distance  # noqa (API import)
+#from datashader.spatial.proximity import manhattan_distance  # noqa (API import)
+#
+##from datashader.spatial.viewshed import viewshed  # noqa (API import)
+#from datashader.spatial.zonal import zonal_stats  # noqa (API import)