--- a/datashader/tests/test_proximity.py
+++ b/datashader/tests/test_proximity.py
@@ -1,7 +1,7 @@
 import pytest
 
-from datashader.spatial import proximity
-from datashader.spatial import great_circle_distance
+from datashader.spatial.proximity import proximity
+from datashader.spatial.proximity import great_circle_distance
 import datashader as ds
 
 import numpy as np