--- a/examples/user_guide/8_Geography.ipynb
+++ b/examples/user_guide/8_Geography.ipynb
@@ -381,7 +381,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "from datashader.spatial import proximity\n",
+    "from datashader.spatial.proximity import proximity\n",
     "from datashader.transfer_functions import dynspread\n",
     "from datashader.transfer_functions import set_background\n",
     "\n",
@@ -510,8 +510,8 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "from datashader.spatial import proximity\n",
-    "from datashader.spatial import viewshed\n",
+    "from datashader.spatial.proximity import proximity\n",
+    "from datashader.spatial.viewshed import viewshed\n",
     "\n",
     "from datashader.transfer_functions import dynspread\n",
     "from datashader.transfer_functions import set_background\n",
@@ -582,7 +582,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "from datashader.spatial import viewshed\n",
+    "from datashader.spatial.viewshed import viewshed\n",
     "\n",
     "W = 600\n",
     "H = 400\n",
@@ -680,7 +680,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "from datashader.spatial import zonal_stats\n",
+    "from datashader.spatial.zonal import zonal_stats\n",
     "\n",
     "zonal_stats(zones_agg, terrain)"
    ]
@@ -719,7 +719,7 @@
    ]
   }
  ],
- "metadata": {
+  "metadata": {
   "language_info": {
    "name": "python",
    "pygments_lexer": "ipython3"