fixes #846 

- [x] fix datashader import slowness...
- [x] Quick fix is to change viewshed, proximity, zonal namespaces: 
```
OLD: from datashader.spatial import viewshed
NEW: from datashader.spatial.viewshed import viewshed
```
# PROFILING ON MASTER:
```
17234270 function calls (16242030 primitive calls) in 17.234 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
       83    0.003    0.000   34.830    0.420 __init__.py:1(<module>)
   1617/1    0.016    0.000   17.244   17.244 {built-in method builtins.exec}
    743/1    0.002    0.000   17.244   17.244 {built-in method builtins.__import__}
   1366/1    0.012    0.000   17.244   17.244 <frozen importlib._bootstrap>:978(_find_and_load)
   1360/1    0.008    0.000   17.243   17.243 <frozen importlib._bootstrap>:948(_find_and_load_unlocked)
   1269/1    0.009    0.000   17.243   17.243 <frozen importlib._bootstrap>:663(_load_unlocked)
   1155/1    0.007    0.000   17.242   17.242 <frozen importlib._bootstrap_external>:722(exec_module)
   1718/1    0.002    0.000   17.242   17.242 <frozen importlib._bootstrap>:211(_call_with_frames_removed)
       16    0.001    0.000   16.117    1.007 core.py:1(<module>)
```

# PROFILING ON THIS PR:
```
2934400 function calls (2811439 primitive calls) in 4.158 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
       83    0.003    0.000    8.722    0.105 __init__.py:1(<module>)
   1614/1    0.016    0.000    4.162    4.162 {built-in method builtins.exec}
    743/1    0.002    0.000    4.162    4.162 {built-in method builtins.__import__}
   1363/1    0.021    0.000    4.162    4.162 <frozen importlib._bootstrap>:978(_find_and_load)
   1357/1    0.008    0.000    4.162    4.162 <frozen importlib._bootstrap>:948(_find_and_load_unlocked)
   1266/1    0.008    0.000    4.160    4.160 <frozen importlib._bootstrap>:663(_load_unlocked)
   1152/1    0.006    0.000    4.160    4.160 <frozen importlib._bootstrap_external>:722(exec_module)
   1715/1    0.002    0.000    4.160    4.160 <frozen importlib._bootstrap>:211(_call_with_frames_removed)
       16    0.001    0.000    3.374    0.211 core.py:
```